Avoid occupation functions as default arguments
Default values are evaluated during function definition. If they originate from
a different module, this module must already be initialized. The current import
strategy ensures this, but if a simple `import elphmod` was used in combination
with importing the modules into the package namespace in the `__init__.py` file,
this could lead to circular-dependency errors (precisely if the required module
was imported later). Such default values are thus better avoided altogether.

The previously used relative imports also ensured timely initialization of the
required module, but they have been dropped for the sake of separating package
and module namespaces.